chore: update props and types
